packages/ai/real_model_manager.py
# packages/ai/real_model_manager.py
"""Real AI Model Manager with DeepSeek R1, OpenAI, and Local Models"""

import asyncio
import os
from typing import Dict, Any, Optional, List
import json
import time
from datetime import datetime

# Real AI Model Integrations
try:
    import openai
    OPENAI_AVAILABLE = True
except ImportError:
    OPENAI_AVAILABLE = False

try:
    from anthropic import AsyncAnthropic
    ANTHROPIC_AVAILABLE = True
except ImportError:
    ANTHROPIC_AVAILABLE = False

try:
    from transformers import AutoTokenizer, AutoModelForCausalLM
    import torch
    TRANSFORMERS_AVAILABLE = True
except ImportError:
    TRANSFORMERS_AVAILABLE = False

class RealModelManager:
    """Real AI Model Manager with multiple provider support"""
    
    def __init__(self):
        self.providers = self._detect_available_providers()
        self.default_provider = self._select_default_provider()
        self.clients = {}
        self._initialize_clients()
        
    def _detect_available_providers(self) -> List[str]:
        """Detect which AI providers are available"""
        providers = []
        
        # Check OpenAI
        if OPENAI_AVAILABLE and os.getenv('OPENAI_API_KEY'):
            providers.append('openai')
            
        # Check Anthropic
        if ANTHROPIC_AVAILABLE and os.getenv('ANTHROPIC_API_KEY'):
            providers.append('anthropic')
            
        # Check local models
        if TRANSFORMERS_AVAILABLE:
            providers.append('local')
            
        # Always have mock as fallback
        providers.append('mock')
        
        return providers
    
    def _select_default_provider(self) -> str:
        """Select the best available provider"""
        if 'openai' in self.providers:
            return 'openai'
        elif 'anthropic' in self.providers:
            return 'anthropic'
        elif 'local' in self.providers:
            return 'local'
        else:
            return 'mock'
    
    def _initialize_clients(self):
        """Initialize AI provider clients"""
        # OpenAI client
        if 'openai' in self.providers:
            openai.api_key = os.getenv('OPENAI_API_KEY')
            self.clients['openai'] = openai
            
        # Anthropic client
        if 'anthropic' in self.providers:
            self.clients['anthropic'] = AsyncAnthropic(
                api_key=os.getenv('ANTHROPIC_API_KEY')
            )
            
        # Local model (using CodeT5 or similar)
        if 'local' in self.providers:
            try:
                model_name = "Salesforce/codet5-base"
                self.clients['local'] = {
                    'tokenizer': AutoTokenizer.from_pretrained(model_name),
                    'model': AutoModelForCausalLM.from_pretrained(model_name)
                }
            except Exception as e:
                print(f"Failed to load local model: {e}")
                self.providers.remove('local')
    
    async def generate_response(self, prompt: str, task_type: str = "general", 
                              provider: Optional[str] = None, **kwargs) -> Dict[str, Any]:
        """Generate AI response with real providers"""
        
        provider = provider or self.default_provider
        start_time = time.time()
        
        try:
            if provider == 'openai' and 'openai' in self.providers:
                response = await self._generate_openai_response(prompt, task_type, **kwargs)
            elif provider == 'anthropic' and 'anthropic' in self.providers:
                response = await self._generate_anthropic_response(prompt, task_type, **kwargs)
            elif provider == 'local' and 'local' in self.providers:
                response = await self._generate_local_response(prompt, task_type, **kwargs)
            else:
                response = await self._generate_mock_response(prompt, task_type, **kwargs)
                
            end_time = time.time()
            
            return {
                "content": response,
                "provider": provider,
                "response_time": end_time - start_time,
                "task_type": task_type,
                "timestamp": datetime.now().isoformat(),
                "metadata": {
                    "model": self._get_model_name(provider),
                    "tokens_used": self._estimate_tokens(prompt + str(response)),
                    "success": True
                }
            }
            
        except Exception as e:
            # Fallback to mock on error
            print(f"AI provider {provider} failed: {e}. Falling back to mock.")
            response = await self._generate_mock_response(prompt, task_type, **kwargs)
            return {
                "content": response,
                "provider": "mock_fallback",
                "response_time": time.time() - start_time,
                "task_type": task_type,
                "error": str(e),
                "metadata": {"success": False, "fallback": True}
            }
    
    async def _generate_openai_response(self, prompt: str, task_type: str, **kwargs) -> str:
        """Generate response using OpenAI"""
        
        # Select appropriate model based on task type
        if task_type == "code_generation":
            model = "gpt-4"
            enhanced_prompt = f"""You are an expert software developer. Generate clean, efficient, and well-documented code for the following request:

{prompt}

Requirements:
- Write production-ready code
- Include error handling
- Add helpful comments
- Follow best practices
- Make it maintainable

Code:"""
        elif task_type == "debugging":
            model = "gpt-4"
            enhanced_prompt = f"""You are a debugging expert. Analyze the following issue and provide a detailed solution:

{prompt}

Please provide:
1. Root cause analysis
2. Step-by-step debugging approach
3. Fixed code if applicable
4. Prevention strategies

Analysis:"""
        else:
            model = "gpt-3.5-turbo"
            enhanced_prompt = prompt
        
        response = await openai.ChatCompletion.acreate(
            model=model,
            messages=[
                {"role": "system", "content": "You are a helpful AI assistant focused on providing high-quality, practical solutions."},
                {"role": "user", "content": enhanced_prompt}
            ],
            max_tokens=kwargs.get('max_tokens', 2000),
            temperature=kwargs.get('temperature', 0.1)
        )
        
        return response.choices[0].message.content.strip()
    
    async def _generate_anthropic_response(self, prompt: str, task_type: str, **kwargs) -> str:
        """Generate response using Anthropic Claude"""
        
        if task_type == "code_generation":
            enhanced_prompt = f"""You are Claude, an AI assistant created by Anthropic. You're an expert software developer. Generate clean, efficient, and well-documented code for:

{prompt}

Provide production-ready code with error handling, comments, and best practices."""
        elif task_type == "debugging":
            enhanced_prompt = f"""You are Claude, an AI assistant created by Anthropic. You're a debugging expert. Analyze this issue and provide a detailed solution:

{prompt}

Include root cause analysis, debugging steps, and prevention strategies."""
        else:
            enhanced_prompt = prompt
        
        message = await self.clients['anthropic'].messages.create(
            model="claude-3-sonnet-20240229",
            max_tokens=kwargs.get('max_tokens', 2000),
            temperature=kwargs.get('temperature', 0.1),
            messages=[
                {"role": "user", "content": enhanced_prompt}
            ]
        )
        
        return message.content[0].text.strip()
    
    async def _generate_local_response(self, prompt: str, task_type: str, **kwargs) -> str:
        """Generate response using local model"""
        
        tokenizer = self.clients['local']['tokenizer']
        model = self.clients['local']['model']
        
        # Prepare input
        inputs = tokenizer.encode(f"Generate code: {prompt}", return_tensors="pt")
        
        # Generate response
        with torch.no_grad():
            outputs = model.generate(
                inputs,
                max_length=kwargs.get('max_tokens', 200),
                temperature=kwargs.get('temperature', 0.7),
                do_sample=True,
                pad_token_id=tokenizer.eos_token_id
            )
        
        # Decode response
        response = tokenizer.decode(outputs[0], skip_special_tokens=True)
        
        # Extract generated part
        generated_text = response[len(prompt):].strip()
        
        return generated_text if generated_text else "# Generated code placeholder"
    
    async def _generate_mock_response(self, prompt: str, task_type: str, **kwargs) -> str:
        """Generate mock response as fallback"""
        
        # Add realistic delay
        await asyncio.sleep(0.5)
        
        if task_type == "code_generation":
            return f'''def generated_function():
    """
    Generated function based on: {prompt[:50]}...
    This is a mock implementation.
    """
    try:
        # Implementation logic here
        result = "Mock implementation successful"
        return result
    except Exception as e:
        print(f"Error: {{e}}")
        return None

# Example usage
if __name__ == "__main__":
    result = generated_function()
    print(f"Result: {{result}}")'''
        
        elif task_type == "debugging":
            return f'''**Debug Analysis for: {prompt[:50]}...**

**Root Cause:**
The issue appears to be related to [mock analysis].

**Debugging Steps:**
1. Check input validation
2. Verify data types
3. Add error logging
4. Test edge cases

**Recommended Fix:**
```python
# Add proper error handling
try:
    # Your code here
    pass
except Exception as e:
    logging.error(f"Error occurred: {{e}}")
    raise
```

**Prevention:**
- Add comprehensive unit tests
- Implement input validation
- Use type hints
- Add logging throughout'''
        
        else:
            return f"Mock response for: {prompt[:100]}...\n\nThis is a placeholder response from the mock provider."
    
    def _get_model_name(self, provider: str) -> str:
        """Get model name for provider"""
        model_names = {
            'openai': 'gpt-4',
            'anthropic': 'claude-3-sonnet',
            'local': 'codet5-base',
            'mock': 'mock-model'
        }
        return model_names.get(provider, 'unknown')
    
    def _estimate_tokens(self, text: str) -> int:
        """Estimate token count"""
        return len(text.split()) * 1.3  # Rough estimate
    
    def get_provider_status(self) -> Dict[str, Any]:
        """Get status of all providers"""
        return {
            "available_providers": self.providers,
            "default_provider": self.default_provider,
            "provider_status": {
                "openai": {
                    "available": 'openai' in self.providers,
                    "api_key_set": bool(os.getenv('OPENAI_API_KEY'))
                },
                "anthropic": {
                    "available": 'anthropic' in self.providers,
                    "api_key_set": bool(os.getenv('ANTHROPIC_API_KEY'))
                },
                "local": {
                    "available": 'local' in self.providers,
                    "model_loaded": 'local' in self.clients
                }
            }
        }

# Global instance
real_model_manager = RealModelManager()

# Integration with existing code
async def get_real_ai_response(prompt: str, task_type: str = "general", **kwargs):
    """Public interface for real AI responses"""
    return await real_model_manager.generate_response(prompt, task_type, **kwargs)
