pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "revoagent"
version = "1.0.0"
description = "Revolutionary Agentic Coding System Platform"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "reVoAgent Team", email = "team@revoagent.dev"}
]
maintainers = [
    {name = "reVoAgent Team", email = "team@revoagent.dev"}
]
keywords = [
    "ai", "agents", "coding", "automation", "software-engineering",
    "browser-automation", "llm", "local-models", "zero-cost"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.12"
dependencies = [
    # Core dependencies
    "pydantic>=2.11.0",
    "pyyaml>=6.0",
    "click>=8.2.0",
    "rich>=14.0.0",
    "asyncio-mqtt>=0.16.0",
    
    # AI/ML dependencies
    "torch>=2.7.0",
    "transformers>=4.52.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "sentence-transformers>=4.1.0",
    "onnxruntime>=1.22.0",
    
    # Web and browser automation
    "playwright>=1.52.0",
    "selenium>=4.33.0",
    "beautifulsoup4>=4.13.0",
    "requests>=2.32.0",
    "httpx>=0.28.0",
    
    # Development tools
    "gitpython>=3.1.40",
    "docker>=7.1.0",
    "psutil>=7.0.0",
    
    # Database and storage
    "redis>=6.2.0",
    
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "websockets>=15.0",
    
    # Utilities
    "python-dotenv>=1.1.0",
    "tenacity>=9.1.0",
    "jinja2>=3.1.0",
    "aiofiles>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.9.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.6.0",
    "pre-commit>=3.5.0",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.4.0",
    "mkdocstrings[python]>=0.23.0",
]
all = [
    "revoagent[dev,docs]"
]

[project.urls]
Homepage = "https://github.com/heinzstkdev/reVoAgent"
Documentation = "https://docs.revoagent.dev"
Repository = "https://github.com/heinzstkdev/reVoAgent"
"Bug Tracker" = "https://github.com/heinzstkdev/reVoAgent/issues"

[project.scripts]
revoagent = "revoagent.cli:main"
revoagent-server = "revoagent.web_server:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["revoagent"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=revoagent",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["src/revoagent"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]